Split payloads so they don't try and read .env
